Catalog facets part 2 (#2111)
* Refactor a bit. Create a utils file.

* Remove unused function.

* Fetch catalog index when enabling it. Fix some linting errors and warnings
