+++ src/components/UI/Display/Apps/AppDetailNew/AppDetailPage.stories.tsx
@@ -13,13 +13,15 @@ const Template: Story<IAppDetailPageProps> = (args) => (
   <AppDetailPage {...args} />
 );
 
+const THEYEAR = 2021;
+
 export const WithReadme = Template.bind({});
 WithReadme.args = {
   appTitle: 'efk-stack-app',
   appIconURL: 'https://dummyimage.com/125/125/fff',
   catalogName: 'Giant Swarm Managed',
   chartVersion: 'v0.3.2',
-  createDate: new Date(2021, 0, 1),
+  createDate: new Date(THEYEAR, 0, 1),
   includesVersion: 'v1.9.0',
   description: 'Open Distro for ElasticSearch',
   website: 'github.com/giantswarm/efk-stack-app',
@@ -33,7 +35,7 @@ WithoutReadme.args = {
   appIconURL: 'https://dummyimage.com/125/125/fff',
   catalogName: 'Giant Swarm Managed',
   chartVersion: 'v0.3.2',
-  createDate: new Date(2021, 0, 1),
+  createDate: new Date(THEYEAR, 0, 1),
   includesVersion: 'v1.9.0',
   description: 'Open Distro for ElasticSearch',
   website: 'github.com/giantswarm/efk-stack-app',
